+++ src/Orchestration/Orchestration.php
@@ -210,11 +210,11 @@ class Orchestration
     }
 
     /**
-     * Set max allowed CPU cores per container
+     * Set max allowed CPU Quota per container
      *
      * @return $this
      */
-    public function setCpus(int $cores): self
+    public function setCpus(float $cores): self
     {
         $this->adapter->setCpus($cores);
